* Fix - removed not used import. * Fix - one test moved into isolated Test file as assertion from console done.
