+++ testApi/src/main/scala/africa/absa/testing/scapi/rest/response/Response.scala
@@ -16,7 +16,7 @@
 
 package africa.absa.testing.scapi.rest.response
 
-import africa.absa.testing.scapi.json.{Environment, ResponseAction}
+import africa.absa.testing.scapi.json.ResponseAction
 import africa.absa.testing.scapi.logging.Logger
 import africa.absa.testing.scapi.rest.model.CookieValue
 import africa.absa.testing.scapi.rest.response.action.types.ResponseActionGroupType
